# internal/handler/auth.go
package handler

import (
	"encoding/json"
	"net/http"

	"pos-multi-branch/backend/internal/config"
	"pos-multi-branch/backend/internal/middleware"
	"pos-multi-branch/backend/internal/model"
	"pos-multi-branch/backend/internal/repository"
)

type AuthHandler struct {
	cfg *config.Config
}

func NewAuthHandler(cfg *config.Config) *AuthHandler {
	return &AuthHandler{cfg: cfg}
}

// Login godoc
// @Summary      User login
// @Description  Authenticate a user with username and password, returns JWT token
// @Tags         Auth
// @Accept       json
// @Produce      json
// @Param request body model.LoginRequest true "Login credentials"
// @Success      200  {object}  model.LoginResponse
// @Failure      400  {object}  map[string]string
// @Failure      401  {object}  map[string]string
// @Failure      500  {object}  map[string]string
// @Router       /auth/login [post]
func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
	var req model.LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid request body"})
		return
	}
	if req.Username == "" || req.Password == "" {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "username and password required"})
		return
	}

	user, err := repository.FindUserByUsername(r.Context(), req.Username)
	if err != nil {
		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "internal error"})
		return
	}
	if user == nil || !repository.VerifyPassword(user.Password, req.Password) {
		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "invalid credentials"})
		return
	}

	token, err := middleware.GenerateToken(user.ID, user.Role, user.BranchID, h.cfg.JWTExpiryHours)
	if err != nil {
		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "token generation failed"})
		return
	}

	writeJSON(w, http.StatusOK, model.LoginResponse{
		Token: token,
		User:  *user,
	})
}

// Me godoc
// @Summary      Get current user
// @Description  Return the authenticated user's profile
// @Tags         Auth
// @Produce      json
// @Success      200  {object}  model.MeResponse
// @Failure      401  {object}  map[string]string
// @Security     BearerAuth
// @Router       /auth/me [get]
func (h *AuthHandler) Me(w http.ResponseWriter, r *http.Request) {
	userID := middleware.GetUserID(r.Context())
	user, err := repository.FindUserByID(r.Context(), userID)
	if err != nil || user == nil {
		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "user not found"})
		return
	}
	writeJSON(w, http.StatusOK, model.MeResponse{User: *user})
}

// Logout godoc
// @Summary      Logout
// @Description  Invalidate current session (stateless JWT — client must discard token)
// @Tags         Auth
// @Produce      json
// @Success      200  {object}  map[string]string
// @Security     BearerAuth
// @Router       /auth/logout [post]
func (h *AuthHandler) Logout(w http.ResponseWriter, r *http.Request) {
	// Stateless JWT — client must discard token.
	writeJSON(w, http.StatusOK, map[string]string{"message": "logged out"})
}